MB-15910: Fix potential segfaults on WIN32 due to incorrect log format strings
Fix mismatch between LOG() formatting strings and the arguments. The
common pattern here is format strings which:

1) contain a char* (%s) formatter
2) preceded by a mis-sized integer formatter.

This results in printf() on 32bit Windows reading too much (or too
little) from the stack for the integer argument; meaning that when it
attempts to handle the following %s formatter and read the char* it
dereferences an invalid pointer.

Reviewed-on: http://review.couchbase.org/53924
Tested-by: buildbot <build@couchbase.com>

+++ src/flusher.cc
@@ -142,8 +142,9 @@ void Flusher::schedule_UNLOCKED() {
 void Flusher::start() {
     LockHolder lh(taskMutex);
     if (taskId) {
-        LOG(EXTENSION_LOG_WARNING, "Double start in flusher task id %llu: %s",
-                taskId, stateName());
+        LOG(EXTENSION_LOG_WARNING,
+            "Double start in flusher task id %" PRIu64 ": %s",
+            uint64_t(taskId), stateName());
         return;
     }
     schedule_UNLOCKED();
